refactor: project folder structure
