+++ .eslintrc.json
@@ -5,6 +5,7 @@
   "extends": ["airbnb", "prettier"],
   "plugins": ["prettier"],
   "rules": {
-    "prettier/prettier": ["error"]
+    "prettier/prettier": ["error"],
+    "no-console": ["off"]
   }
 }
