+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "iot-manager-broker",
-  "version": "0.0.0",
-  "private": true,
+  "version": "0.0.1",
   "scripts": {
     "start": "node ./bin/www",
     "lint:js": "eslint --ext .js . --ignore-pattern \"node_modules/\"",
