--- routes/index.js
@@ -1,8 +1,0 @@
-const express = require('express');
-
-const router = express.Router();
-
-/* GET home page. */
-router.get('/', (req, res) => res.render('index', { title: 'Express TESTING' }));
-
-module.exports = router;
