+++ src/utils.js
@@ -1,0 +1,12 @@
+function parseMessage(message) {
+  try {
+    const item = JSON.parse(message);
+    return item;
+  } catch (e) {
+    return message.toString();
+  }
+}
+
+module.exports = {
+  parseMessage,
+};
